arranged components and container based on new routes
